비밀번호 로직 추가

+++ src/main/java/org/example/newscheduleproject/service/UserService.java
@@ -1,15 +1,12 @@
 package org.example.newscheduleproject.service;
 
-
-
-
 import lombok.RequiredArgsConstructor;
 import org.example.newscheduleproject.dto.Request.UserRequest;
 import org.example.newscheduleproject.dto.Response.UserResponse;
 import org.example.newscheduleproject.entity.User;
 import org.example.newscheduleproject.repository.UserRepository;
 import org.springframework.stereotype.Service;
-
+import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,14 +15,16 @@ import java.util.List;
 
 public class UserService {
     private final UserRepository userRepository;
-    
+
+
+    @Transactional
     public UserResponse saveUser(UserRequest userRequest) {
-        User user = new User(userRequest.getName(), userRequest.getEmail());
+        User user = new User(userRequest.getName(), userRequest.getEmail(),userRequest.getPassword());
         userRepository.save(user);
 
         return UserResponse.from(user);
     }
-
+    @Transactional
     public List<UserResponse> getUser() {
         List<User> users = userRepository.findAll();
         List<UserResponse> userResponseList = new ArrayList<>();
@@ -34,22 +33,29 @@ public class UserService {
         }
         return userResponseList;
     }
-
+    @Transactional
     public UserResponse Uqdate(Long userId, UserRequest userRequest) {
         User user = userRepository.findById(userId).orElseThrow(
                 () -> new IllegalArgumentException("그런 일정이 없는데?")
         );
+        if (!userRequest.getPassword().equals(user.getPassword())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않다");
+        }
 
         user.update(userRequest.getName(),userRequest.getEmail());
         userRepository.save(user);
         return UserResponse.from(user);
-    
-    }
 
-    public void deleteUserById(Long userId) {
+    }
+    //삭제
+    @Transactional
+    public void deleteUserById(Long userId,UserRequest userRequest) {
         User user = userRepository.findById(userId).orElseThrow(
                 () -> new IllegalArgumentException("그런 아이디 없는데요?")
         );
+        if (!userRequest.getPassword().equals(user.getPassword())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않다");
+        }
         userRepository.delete(user);
     }
 }
